+++ src/grocker/utils.py
@@ -108,11 +108,9 @@ def parse_config(config_paths, **kwargs):
 
     for config_path in config_paths:
         project_config = helpers.load_yaml(config_path)
-        config.update(project_config or {})
+        config = helpers.deep_update(config, project_config or {})
 
-    config.update({k: v for k, v in kwargs.items() if v})
-
-    return config
+    return helpers.deep_update(config, {k: v for k, v in kwargs.items() if v})
 
 
 class GrockerRequirement: